Poprawione cast cos tam

+++ src/main/java/Product.java
@@ -29,7 +29,7 @@ public class Product {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
